brainstorm_agent_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
redis==5.0.1
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
asyncpg==0.28.0
aiofiles==23.2.1
python-multipart==0.0.6
